VidUtils/garbage/requirements.txt
ffmpeg-python==0.2.0
boto3==1.34.34
openai==1.11.0
praw==7.7.1
python-dotenv==1.0.0
requests==2.31.0
update-checker==0.18.0
redis==4.6.0
rq==1.15.1
pydantic==1.10.8
elevenlabs==0.2.15
google-cloud-texttospeech==2.14.1
google-cloud-storage==2.13.0
beautifulsoup4==4.12.2
lxml==4.9.2
numpy==1.25.0
aeneas==1.7.3.0
pydub==0.25.1
pillow==10.2.0
pilmoji==2.0.4
